requirements.txt
# Core dependencies
Flask==3.0.2
python-dotenv==1.0.0
pyyaml==6.0.1

# Web scraping and networking
undetected-chromedriver==3.5.5
selenium==4.12.0
beautifulsoup4==4.12.2
requests==2.31.0
urllib3>=1.25.4,<1.27  # Fixed version range for compatibility

# Image and PDF processing
Pillow==10.4.0
img2pdf==0.4.4

# Storage
boto3==1.28.44
botocore==1.31.44

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Type stubs
types-requests==2.31.0.2
types-PyYAML==6.0.12.11
types-beautifulsoup4==4.12.0.6
